Work on the gameplay scene
Added the Cinemachine and Input system packages from Unity with the intent to have a CM camera follow the player using the new input system so people can play with mouse & keyboard or gamepads. None of it work currently but its stubbed out. Also created some programmer art assets to use as a template for starting to design. NOT FINAL ART.